+++ src/commands/submit.rs
@@ -34,7 +34,7 @@ pub struct OptSubmit {
     pub testcases: Option<Vec<String>>,
 
     /// Display limit for the test
-    #[structopt(long, short("l"), value_name("SIZE"), default_value("4KiB"))]
+    #[structopt(long, value_name("SIZE"), default_value("4KiB"))]
     pub display_limit: Size,
 
     /// Path to `snowchains.dhall`